+++ src/main/java/insane96mcp/enhancedai/ai/EAAvoidEntityGoal.java
@@ -51,8 +51,8 @@ public class EAAvoidEntityGoal<T extends LivingEntity> extends Goal {
 		this.goalOwner = entityIn;
 		this.classToAvoid = avoidClass;
 		this.avoidTargetSelector = targetPredicate;
-		this.avoidDistance = avoidDistance * avoidDistance;
-		this.avoidDistanceNear = avoidDistanceNear * avoidDistanceNear;
+		this.avoidDistance = avoidDistance;
+		this.avoidDistanceNear = avoidDistanceNear;
 		this.nearSpeed = nearSpeedIn;
 		this.farSpeed = farSpeedIn;
 		this.predicateOnAvoidEntity = predicate;
@@ -60,7 +60,7 @@ public class EAAvoidEntityGoal<T extends LivingEntity> extends Goal {
 	}
 
 	public boolean canUse() {
-		this.avoidTarget = this.goalOwner.level().getNearestEntity(this.classToAvoid, this.builtTargetSelector, this.goalOwner, this.goalOwner.getX(), this.goalOwner.getY(), this.goalOwner.getZ(), this.goalOwner.getBoundingBox().inflate(this.avoidDistance, this.avoidDistance, this.avoidDistance));
+		this.avoidTarget = this.goalOwner.level().getNearestEntity(this.classToAvoid, this.builtTargetSelector, this.goalOwner, this.goalOwner.getX(), this.goalOwner.getY(), this.goalOwner.getZ(), this.goalOwner.getBoundingBox().inflate(this.avoidDistance));
 		if (this.avoidTarget == null) {
 			return false;
 		} else {
@@ -89,7 +89,7 @@ public class EAAvoidEntityGoal<T extends LivingEntity> extends Goal {
 	}
 
 	public void tick() {
-		if (this.goalOwner.distanceToSqr(this.avoidTarget) < this.avoidDistanceNear) {
+		if (this.goalOwner.distanceToSqr(this.avoidTarget) < this.avoidDistanceNear * this.avoidDistanceNear) {
 			this.goalOwner.getNavigation().setSpeedModifier(this.nearSpeed);
 		} else {
 			this.goalOwner.getNavigation().setSpeedModifier(this.farSpeed);